chore: rename back to actual directory names
