+++ client/components/page/Admin/Invoice/InvoiceList.page.tsx
@@ -1,5 +1,5 @@
 import BannerCover from '@/components/layout/BannerCover';
-import { Footer } from '@/components/shared/footertmp';
+import { Footer } from '@/components/shared/Footer';
 import { SidebarWithHeader } from '@/components/shared/Header';
 import { InvoiceListNextPageProps } from 'pages/admin/invoice/';
 import InvoiceList from './InvoiceList';
