+++ client/components/page/Admin/PlanOption/PlanOptionList.page.tsx
@@ -1,5 +1,5 @@
 import BannerCover from '@/components/layout/BannerCover';
-import { Footer } from '@/components/shared/footertmp';
+import { Footer } from '@/components/shared/Footer';
 import { SidebarWithHeader } from '@/components/shared/Header';
 import { PlanOptionListNextPageProps } from 'pages/admin/planOption/';
 import PlanOptionList from './PlanOptionList';
