+++ client/components/page/Admin/Product/ProductDetail.page.tsx
@@ -1,5 +1,5 @@
 import BannerCover from '@/components/layout/BannerCover';
-import { Footer } from '@/components/shared/footertmp';
+import { Footer } from '@/components/shared/Footer';
 import { SidebarWithHeader } from '@/components/shared/Header';
 import { ProductDetailNextPageProps } from 'pages/admin/product/[productId]';
 import ProductDetail from './ProductDetail';
