+++ client/components/page/Admin/Product/ProductList.page.tsx
@@ -1,5 +1,5 @@
 import BannerCover from '@/components/layout/BannerCover';
-import { Footer } from '@/components/shared/footertmp';
+import { Footer } from '@/components/shared/Footer';
 import { SidebarWithHeader } from '@/components/shared/Header';
 import { ProductListNextPageProps } from 'pages/admin/product/';
 import ProductList from './ProductList';
