+++ client/components/page/Admin/Promotion/PromotionList.page.tsx
@@ -1,5 +1,5 @@
 import BannerCover from '@/components/layout/BannerCover';
-import { Footer } from '@/components/shared/footertmp';
+import { Footer } from '@/components/shared/Footer';
 import { SidebarWithHeader } from '@/components/shared/Header';
 import { PromotionListNextPageProps } from 'pages/admin/promotion/';
 import PromotionList from './PromotionList';
