+++ client/components/shared/Footer/index.ts
